# docs/swagger.yaml
basePath: /swagger/datahub-service
definitions:
  controllers.APIResponse:
    properties:
      data: {}
      msg:
        example: 操作成功
        type: string
      status:
        example: 0
        type: integer
    type: object
...
  controllers.PaginatedResponse:
    properties:
      data: {}
      msg:
        example: 操作成功
        type: string
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      status:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  models.DataMaskingRule:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      creator_id:
        type: string
      creator_name:
        type: string
      data_source:
        type: string
      data_table:
        type: string
      field_name:
        type: string
      field_type:
        type: string
      id:
        type: string
      is_enabled:
        type: boolean
      masking_config:
        additionalProperties: true
        type: object
      masking_type:
        description: mask/replace/encrypt/pseudonymize
        type: string
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.Metadata:
    properties:
      content:
        additionalProperties: true
        type: object
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      name:
        type: string
      related_object_id:
        type: string
      related_object_type:
        description: basic_library/data_interface/thematic_library等
        type: string
      type:
        description: technical/business/management
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
  description: 智慧园区数据底座后台服务，提供数据采集、处理、存储、治理和共享功能
  title: 数据底座服务 API
  version: "1.0"
paths:
...
  /governance/masking-rules:
    get:
      consumes:
      - application/json
      description: 分页获取数据脱敏规则列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 数据源
        in: query
        name: data_source
        type: string
      - description: 脱敏类型
        in: query
        name: masking_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DataMaskingRule'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取数据脱敏规则列表
      tags:
      - 数据治理
    post:
      consumes:
      - application/json
      description: 创建新的数据脱敏规则
      parameters:
      - description: 数据脱敏规则信息
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.DataMaskingRule'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.DataMaskingRule'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建数据脱敏规则
      tags:
      - 数据治理
  /governance/masking-rules/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的数据脱敏规则
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除数据脱敏规则
      tags:
      - 数据治理
    get:
      consumes:
      - application/json
      description: 根据ID获取数据脱敏规则详情
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.DataMaskingRule'
              type: object
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 根据ID获取数据脱敏规则
      tags:
      - 数据治理
    put:
      consumes:
      - application/json
      description: 更新数据脱敏规则信息
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 更新数据脱敏规则
      tags:
      - 数据治理
  /governance/metadata:
    get:
      consumes:
      - application/json
      description: 分页获取元数据列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 元数据类型
        in: query
        name: type
        type: string
      - description: 元数据名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Metadata'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取元数据列表
      tags:
      - 数据治理
    post:
      consumes:
      - application/json
      description: 创建新的元数据
      parameters:
      - description: 元数据信息
        in: body
        name: metadata
        required: true
        schema:
          $ref: '#/definitions/models.Metadata'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Metadata'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建元数据
      tags:
      - 数据治理
  /governance/metadata/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的元数据
      parameters:
      - description: 元数据ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 元数据不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 删除元数据
      tags:
      - 数据治理
    get:
      consumes:
      - application/json
      description: 根据ID获取元数据详情
      parameters:
      - description: 元数据ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Metadata'
              type: object
        "404":
          description: 元数据不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 根据ID获取元数据
      tags:
      - 数据治理
    put:
      consumes:
      - application/json
      description: 更新元数据信息
      parameters:
      - description: 元数据ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "404":
          description: 元数据不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 更新元数据
      tags:
      - 数据治理
  /governance/quality-check:
    post:
      consumes:
      - application/json
      description: 对指定对象执行数据质量检查并生成报告
      parameters:
      - description: 对象ID
        in: query
        name: object_id
        required: true
        type: string
      - description: 对象类型
        in: query
        name: object_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 检查成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.DataQualityReport'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 执行数据质量检查
      tags:
      - 数据治理
  /governance/quality-reports:
    get:
      consumes:
      - application/json
      description: 分页获取数据质量报告列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 对象类型
        in: query
        name: object_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DataQualityReport'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取数据质量报告列表
      tags:
      - 数据治理
  /governance/quality-reports/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取数据质量报告详情
      parameters:
      - description: 报告ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.DataQualityReport'
              type: object
        "404":
          description: 报告不存在
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 根据ID获取数据质量报告
      tags:
      - 数据治理
  /governance/quality-rules:
    get:
      consumes:
      - application/json
      description: 分页获取数据质量规则列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 规则类型
        in: query
        name: type
        type: string
      - description: 关联对象类型
        in: query
        name: object_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.QualityRule'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 获取数据质量规则列表
      tags:
      - 数据治理
    post:
      consumes:
      - application/json
      description: 创建新的数据质量规则
      parameters:
      - description: 数据质量规则信息
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.QualityRule'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/controllers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.QualityRule'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.APIResponse'
      summary: 创建数据质量规则
      tags:
      - 数据治理
...
swagger: "2.0"